+++ src/rgw/store/sfs/CHANGELOG.md
@@ -7,6 +7,11 @@ The format is based on
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+- Fixed queries to users by access key when user has multiple keys.
 ## [0.6.0] - 2022-09-29
 
 ### Added